requirements.txt
construct==2.10.68
greenlet==2.0.2
python-magic==0.4.27
PyYAML==6.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
python-memcached==1.59
six==1.16.0
requests==2.28.2
